Remove casting of invalid values
Previously, known boolean values were casted to `true`, and known
numerical values were casted to a number.  Even if those values
were invalid, which resulted in `NaN`.

Now, this is reverted and invalid booleans (not *overloaded* booleans)
are only casted to `true` when they have a valid value (empty or the
attribute’s name), and numeric values are only casted if they result
in a valid number.