Add support for using `h`, `s` as a JSX pragmas
This change tests that `hastscript` can be used as the pragma for JSX with bublé
and babel.

Code-wise, this adds support for using `h` to generate root nodes.
This is done by omitting the tag name (like so: `h()`, `h(null, 'child')`).
Previously, omitting a `name` resulted in the default element to be created
(`div` for `h`, `g` for `s`).

This change thus is a breaking change.
The old behavior is still available when passing an empty string: `h('')`.

Another aspect of supporting JSX is supporting fragments as children.
As fragments yield root nodes, we unravel them and use only their children.
While this could be seen a change, hast prohibits roots occurring in nodes,
so the unraveling instead fixes what would otherwise be a broken tree.

Related to: syntax-tree/xastscript#3.
Related to: syntax-tree/xastscript#4.