Add better custom element support by tightening overload detection
Tighted the disambiguation for `h(tagName, propsOrChild)`, to see `props` as `child`
only when it can never be properties.
For realworld hast use cases, this pratically adds support for properties with `type`
(`string`) *and* `value` on custom elements.
And removes support for literal nodes (those with a `value`) from being passed by
omitting properties.
You likely don’t need to pass literals as the second argument.
As the standard hast `text` literal can always be passed as just the string.
Or wrap it in an array.
Or pass empty props.

Closes GH-21.